refactor(auth): 공통 인증 계약과 가드 구조 정리
