+++ AGENTS.md
@@ -1,16 +1,19 @@
 ## Language Rule
+
 - **Always respond in Korean.**
 - Regardless of the input language, always reply in Korean.
 
 # Repository Guidelines
 
 ## Project Structure & Module Organization
+
 - `src/` contains the NestJS application code (entrypoint: `src/main.ts`, root module: `src/app.module.ts`).
 - `test/` holds end-to-end tests and Jest e2e config (e.g., `test/app.e2e-spec.ts`).
 - `prisma/` contains the Prisma schema and migrations (e.g., `prisma/schema.prisma`).
 - `dist/` is the build output (generated by `pnpm build`).
 
 ## Architecture Rules (Clean Architecture)
+
 - Apply the domain-centered structure for all domains (auth처럼 적용):
   - 기본 레이아웃: `presentation/`, `application/`, `domain/`, `infrastructure/`.
   - 의존성 방향: Presentation → Application → Domain, Infrastructure는 내부 레이어 구현.
@@ -19,8 +22,27 @@
   - `presentation`: controller/guard/strategy 등 요청 입출력 담당.
   - `domain`: entity, repository interface, 타입 정의.
   - `infrastructure`: Prisma 등 외부 구현체.
+- 용어 규칙:
+  - `auth`, `clips`, `users`, `folders`, `workspaces`는 `기능 도메인`으로 부른다.
+  - 각 기능 도메인 내부의 `domain/` 폴더는 `도메인 레이어`로 부른다.
+  - “도메인”이라는 표현만 단독으로 쓰지 말고, 가능한 한 `기능 도메인` / `도메인 레이어`로 구분해서 해석한다.
+- 기능 도메인 경계 규칙:
+  - 다른 기능 도메인의 `presentation/`, `application/`, `domain/`, `infrastructure/`를 직접 import하지 않는다.
+  - 기능 도메인 간에 공유해야 하는 타입/가드/계약은 해당 기능 도메인 내부에 두지 말고 `src/common`으로 이동한다.
+  - 현재 공통 인증 계약은 `src/common/types`와 `src/common/presentation/guards`를 사용한다.
+  - 예: `AuthContext`, `AuthPlatform`, `AuthProvider`, `JwtAccessGuard`는 `auth` 기능 도메인이 아니라 `common`에 둔다.
+- 레이어 배치 기준:
+  - 비즈니스 개념 자체를 표현하는 타입은 각 기능 도메인의 `domain/`에 둔다.
+  - 유스케이스 입출력, 포트 계약, 유스케이스 전용 모델은 각 기능 도메인의 `application/`에 둔다.
+  - HTTP 요청/응답 DTO, Guard, Strategy, Request user 타입처럼 웹 어댑터 성격의 요소는 `presentation/`에 둔다.
+  - 단, 여러 기능 도메인에서 함께 쓰는 웹/인증 계약은 `src/common`으로 올린다.
+- 새 기능을 추가하거나 리팩터링할 때 체크할 것:
+  - 이 import가 다른 기능 도메인의 내부 레이어를 직접 참조하는가?
+  - 여러 기능 도메인에서 재사용되는 타입/가드인데 특정 기능 도메인 내부에 들어가 있지 않은가?
+  - 공통화가 필요하다면 `src/common`으로 옮긴 뒤 각 기능 도메인이 공통 모듈만 참조하게 만든다.
 
 ## Build, Test, and Development Commands
+
 - `pnpm start`: run the app with NestJS in default mode.
 - `pnpm start:dev`: run in watch mode for local development.
 - `pnpm start:debug`: run with the Node inspector attached.
@@ -31,17 +53,18 @@
 - `pnpm test:e2e`: run end-to-end tests with `test/jest-e2e.json`.
 
 ## Coding Style & Naming Conventions
+
 - Language: TypeScript with NestJS conventions (modules, controllers, services).
 - Formatting: Prettier is the source of truth; run `pnpm format` before commits.
 - Linting: ESLint with Prettier integration; fix issues via `pnpm lint`.
 - Tests: use `.spec.ts` for unit tests (e.g., `src/app.controller.spec.ts`).
 
 ## Testing Guidelines
+
 - Framework: Jest with `ts-jest` transform; unit tests live in `src/`.
 - E2E: Jest config in `test/jest-e2e.json`; use `*.e2e-spec.ts` naming.
 - Coverage: optional via `pnpm test:cov` (outputs to `coverage/`).
 - 테스트 범위: 유스케이스에 대한 테스트만 작성한다. 컨트롤러 테스트는 작성하지 않는다.
-
 
 ## Commit & Pull Request Guidelines
 
@@ -205,7 +228,6 @@ When the maintainer requests an issue to be created:
 
 1. **Interpretation**  
    AI reads the user's request and extracts:
-
    - The core goal of the issue or feature request
    - Any cause or background context
    - Implementation plan or improvement ideas (if provided)
@@ -213,7 +235,6 @@ When the maintainer requests an issue to be created:
 
 2. **Classification**  
    AI determines the appropriate issue type:
-
    - Bug → bug_report.yml
    - Feature → feature_implementation.yml
    - Improvement → improvement.yml
@@ -251,7 +272,7 @@ AI must respond to instructions such as:
 - Secrets belong in `.env.local`; never commit environment files. Reference via `process.env` and document required keys in the PR.
 - Keep secrets out of code and logs; document new config keys in the PR.
 
-
 ## Configuration Notes
+
 - Prisma: update `prisma/schema.prisma` and generate migrations in `prisma/migrations/`.
 - Environment: document required variables in the PR description if you add any.
