+++ src/auth/infrastructure/prisma-auth.repository.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@nestjs/common';
+import { AuthProvider } from 'src/common/types/auth-provider.type';
 import {
   AuthProvider as PrismaAuthProvider,
   SubscriptionPlan,
@@ -17,7 +18,6 @@ import {
   CreateAuthAccountInput,
   UserInfo,
 } from '../domain/auth.repository';
-import { AuthProvider } from '../domain/auth.types';
 
 @Injectable()
 export class PrismaAuthRepository implements AuthRepository {
