+++ src/auth/presentation/guards/jwt-refresh-token.guard.ts
@@ -7,8 +7,8 @@ import {
 import { JwtService } from '@nestjs/jwt';
 import { Request } from 'express';
 import { ConfigService } from '@nestjs/config';
-import { AuthContext } from 'src/auth/application/auth-context';
-import { AuthPlatform } from 'src/auth/domain/auth.types';
+import { AuthContext } from 'src/common/types/auth-context.type';
+import { AuthPlatform } from 'src/common/types/auth-platform.type';
 
 type JwtClaims = {
   sub: string;
