+++ src/auth/presentation/strategies/google.strategy.ts
@@ -1,10 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
-import { AuthProvider, OAuthUser } from '../../domain/auth.types';
+import { AuthProvider } from 'src/common/types/auth-provider.type';
 import { Request } from 'express';
 import type { AuthenticateOptions } from 'passport';
 import { Profile, Strategy } from 'passport-google-oauth20';
 import { ConfigService } from '@nestjs/config';
+import { OAuthUser } from '../../domain/auth.types';
 
 @Injectable()
 export class GoogleStrategy extends PassportStrategy(Strategy, 'google') {
