+++ src/common/types/auth-platform.type.ts
@@ -1,0 +1,1 @@
+export type AuthPlatform = 'WEB' | 'APP';
