+++ src/common/types/auth-provider.type.ts
@@ -1,0 +1,6 @@
+export const AuthProvider = {
+  GOOGLE: 'GOOGLE',
+  GITHUB: 'GITHUB',
+} as const;
+
+export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider];
