+++ src/workspaces/presentation/workspaces.controller.ts
@@ -7,8 +7,8 @@ import {
   UseGuards,
   UseFilters,
 } from '@nestjs/common';
-import { AuthContext } from 'src/auth/application/auth-context';
-import { JwtAccessGuard } from 'src/auth/presentation/guards/jwt-access-token.guard';
+import { AuthContext } from 'src/common/types/auth-context.type';
+import { JwtAccessGuard } from 'src/common/presentation/guards/jwt-access.guard';
 import { ApplicationExceptionFilter } from 'src/common/presentation/filters/application-exception.filter';
 import { GetMySubscriptionUseCase } from '../application/usecases/get-my-subscription.usecase';
 import { UpdateMySubscriptionUseCase } from '../application/usecases/update-my-subscription.usecase';
